analyzegeo/requirements.txt
# SEO vs AISO Analysis Project Requirements

# Web Scraping
selenium==4.15.0
undetected-chromedriver==3.5.4
webdriver-manager==4.0.1
beautifulsoup4==4.12.2
lxml==4.9.3
requests==2.31.0

# Data Analysis
pandas==2.1.3
numpy==1.24.4
scipy==1.11.4

# Machine Learning
scikit-learn==1.3.2

# Visualization
matplotlib==3.8.2
seaborn==0.13.0

# Environment Management
python-dotenv==1.0.0

# Text Processing
nltk==3.8.1

# Utilities
tqdm==4.66.1
setuptools==69.0.0

# Development/Debugging (optional)
jupyter==1.0.0
ipython==8.18.1

# For Perplexity API Integration (when available)
openai==1.3.7  # Similar API structure, may be useful for reference